FSQ Support part 2
Adds the ability to select one mode to support at compile time.
This gives us a hope of fitting in the available memory, as there is
no way it will fit in the Arduino Nano with ALL modes (although that
can be selected in case you're running on a bgger MCU).

Adds code to enable configuring the data to send via the CAT interface
and if all modes are anabled, the FSQ Beacon menu option will ask for
the beacon mode as well as the interval.